requirements.txt
astroid==1.6.5
autopep8==1.3.5
aws-cli-utils==3.1.0
awscli==1.15.27
awslogs==0.10.0
beautifulsoup4==4.6.0
boto3==1.7.25
botocore==1.10.27
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
colormath==3.0.0
cycler==0.10.0
decorator==4.3.0
dlib==19.16.0
docopt==0.6.2
docutils==0.14
idna==2.7
imutils==0.5.1
isort==4.3.4
jmespath==0.9.3
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.2.4
matplotlib==2.2.2
mccabe==0.6.1
networkx==2.1
numpy==1.14.3
opencv-python==3.4.3.18
Pillow==5.3.0
pipreqs==0.4.9
psutil==5.4.5
pyasn1==0.4.2
pycodestyle==2.4.0
pylint==1.9.2
pymongo==3.7.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
pywindowx==0.0.1
PyYAML==3.12
requests==2.19.1
rsa==3.4.2
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
sklearn==0.0
termcolor==1.1.0
tqdm==4.23.4
urllib3==1.23
virtualenv==16.0.0
wrapt==1.10.11
xmltodict==0.11.0
yarg==0.1.9
yq==2.6.0
